+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/AnnotationSpacingRule.kt
@@ -81,7 +81,7 @@ public class AnnotationSpacingRule : Rule("annotation-spacing") {
         if (next != null) {
             if (node.elementType != ElementType.FILE_ANNOTATION_LIST && next.isPartOfComment()) {
                 val psi = node.psi
-                emit(psi.endOffset - 1, ERROR_MESSAGE, true)
+                emit(psi.endOffset, ERROR_MESSAGE, true)
                 if (autoCorrect) {
                     // Special-case autocorrection when the annotation is separated from the annotated construct
                     // by a comment: we need to swap the order of the comment and the annotation
@@ -112,7 +112,7 @@ public class AnnotationSpacingRule : Rule("annotation-spacing") {
             // Check to make sure there are multi breaks between annotations
             if (whiteSpaces.any { psi -> psi.textToCharArray().count { it == '\n' } > 1 }) {
                 val psi = node.psi
-                emit(psi.endOffset - 1, ERROR_MESSAGE, true)
+                emit(psi.endOffset, ERROR_MESSAGE, true)
                 if (autoCorrect) {
                     removeIntraLineBreaks(node, annotations.last())
                     removeExtraLineBreaks(node)